+++ internal/provider/resource_cloudflare_pages_project.go
@@ -3,7 +3,11 @@ package provider
 import (
 	"context"
 	"fmt"
+	"reflect"
+	"strings"
 	"time"
+
+	"github.com/hashicorp/terraform-plugin-log/tflog"
 
 	"github.com/MakeNowJust/heredoc/v2"
 	"github.com/cloudflare/cloudflare-go"
@@ -18,6 +22,9 @@ func resourceCloudflarePagesProject() *schema.Resource {
 		ReadContext:   resourceCloudflarePagesProjectRead,
 		UpdateContext: resourceCloudflarePagesProjectUpdate,
 		DeleteContext: resourceCloudflarePagesProjectDelete,
+		Importer: &schema.ResourceImporter{
+			StateContext: resourceCloudflarePagesProjectImport,
+		},
 		Description: heredoc.Doc(`
 			Provides a resource which manages Cloudflare Pages projects.
 		`),
@@ -83,6 +90,46 @@ func buildDeploymentConfig(environment interface{}) cloudflare.PagesProjectDeplo
 	return config
 }
 
+func parseDeployementConfig(deployment cloudflare.PagesProjectDeploymentConfigEnvironment) (returnValue []map[string]interface{}) {
+	config := make(map[string]interface{})
+
+	config["compatibility_date"] = deployment.CompatibilityDate
+	config["compatibility_flags"] = deployment.CompatibilityFlags
+
+	deploymentVars := map[string]string{}
+	for key, value := range deployment.EnvVars {
+		deploymentVars[key] = value.Value
+	}
+	config["environment_variables"] = deploymentVars
+
+	deploymentVars = map[string]string{}
+	for key, value := range deployment.KvNamespaces {
+		deploymentVars[key] = value.Value
+	}
+	config["kv_namespaces"] = deploymentVars
+
+	deploymentVars = map[string]string{}
+	for key, value := range deployment.DoNamespaces {
+		deploymentVars[key] = value.Value
+	}
+	config["durable_object_namespaces"] = deploymentVars
+
+	deploymentVars = map[string]string{}
+	for key, value := range deployment.R2Bindings {
+		deploymentVars[key] = value.Name
+	}
+	config["r2_buckets"] = deploymentVars
+
+	deploymentVars = map[string]string{}
+	for key, value := range deployment.D1Databases {
+		deploymentVars[key] = value.ID
+	}
+	config["d1_databases"] = deploymentVars
+
+	returnValue = append(returnValue, config)
+	return
+}
+
 func buildPagesProject(d *schema.ResourceData) cloudflare.PagesProject {
 	project := cloudflare.PagesProject{}
 	project.Name = d.Get("name").(string)
@@ -102,7 +149,7 @@ func buildPagesProject(d *schema.ResourceData) cloudflare.PagesProject {
 		if webAnalyticsTag, ok := d.GetOk("build_config.0.web_analytics_tag"); ok {
 			buildConfig.WebAnalyticsTag = webAnalyticsTag.(string)
 		}
-		if webAnalyticsToken, ok := d.GetOk("build_config.0.web_analytics_tag"); ok {
+		if webAnalyticsToken, ok := d.GetOk("build_config.0.web_analytics_token"); ok {
 			buildConfig.WebAnalyticsToken = webAnalyticsToken.(string)
 		}
 		project.BuildConfig = buildConfig
@@ -133,6 +180,9 @@ func buildPagesProject(d *schema.ResourceData) cloudflare.PagesProject {
 			if productionBranch, ok := d.GetOk("source.0.config.0.production_branch"); ok {
 				sourceConfig.ProductionBranch = productionBranch.(string)
 			}
+			if productionBranchEnable, ok := d.GetOk("source.0.config.0.production_deployment_enabled"); ok {
+				sourceConfig.ProductionDeploymentsEnabled = productionBranchEnable.(bool)
+			}
 			if previewBranchIncludes, ok := d.GetOk("source.0.config.0.preview_branch_includes"); ok {
 				for _, item := range previewBranchIncludes.([]interface{}) {
 					sourceConfig.PreviewBranchIncludes = append(sourceConfig.PreviewBranchIncludes, item.(string))
@@ -148,6 +198,7 @@ func buildPagesProject(d *schema.ResourceData) cloudflare.PagesProject {
 			}
 			source.Config = &sourceConfig
 		}
+		project.Source = &source
 	}
 
 	if previewConfig, ok := d.GetOk("deployment_configs.0.preview.0"); ok {
@@ -163,17 +214,69 @@ func buildPagesProject(d *schema.ResourceData) cloudflare.PagesProject {
 func resourceCloudflarePagesProjectRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*cloudflare.API)
 	accountID := d.Get("account_id").(string)
-	projectName := d.Get("name").(string)
 
-	res, err := client.PagesProject(ctx, accountID, projectName)
+	project, err := client.PagesProject(ctx, accountID, d.Id())
 	if err != nil {
-		return diag.FromErr(fmt.Errorf("error reading cloudflare pages project %q: %w", projectName, err))
+		return diag.FromErr(fmt.Errorf("error reading cloudflare pages project %q: %w", d.Id(), err))
 	}
 
-	d.SetId(res.ID)
-	d.Set("subdomain", res.SubDomain)
-	d.Set("created_on", res.CreatedOn.Format(time.RFC3339))
-	d.Set("domains", res.Domains)
+	tflog.Debug(ctx, fmt.Sprintf("Cloudflare Pages Project Response: %#v", project))
+	d.Set("subdomain", project.SubDomain)
+	d.Set("production_branch", project.ProductionBranch)
+	d.Set("account_id", accountID)
+	d.Set("domains", project.Domains)
+	d.Set("created_on", project.CreatedOn.Format(time.RFC3339))
+
+	if project.Source != nil {
+		source := []map[string]interface{}{}
+		source = append(source, map[string]interface{}{
+			"type": project.Source.Type,
+			"config": []map[string]interface{}{
+				{
+					"owner":                         project.Source.Config.Owner,
+					"repo_name":                     project.Source.Config.RepoName,
+					"production_branch":             project.Source.Config.ProductionBranch,
+					"pr_comments_enabled":           project.Source.Config.PRCommentsEnabled,
+					"deployments_enabled":           project.Source.Config.DeploymentsEnabled,
+					"production_deployment_enabled": project.Source.Config.ProductionDeploymentsEnabled,
+					"preview_branch_includes":       project.Source.Config.PreviewBranchIncludes,
+					"preview_branch_excludes":       project.Source.Config.PreviewBranchExcludes,
+					"preview_deployment_setting":    project.Source.Config.PreviewDeploymentSetting,
+				},
+			},
+		},
+		)
+		d.Set("source", source)
+	}
+	emptyProjectBuildConfig := cloudflare.PagesProjectBuildConfig{}
+	if project.BuildConfig != emptyProjectBuildConfig {
+		buildConfig := []map[string]interface{}{}
+		buildConfig = append(buildConfig, map[string]interface{}{
+			"build_command":       project.BuildConfig.BuildCommand,
+			"destination_dir":     project.BuildConfig.DestinationDir,
+			"root_dir":            project.BuildConfig.RootDir,
+			"web_analytics_tag":   project.BuildConfig.WebAnalyticsTag,
+			"web_analytics_token": project.BuildConfig.WebAnalyticsToken,
+		},
+		)
+		d.Set("build_config", buildConfig)
+	}
+
+	emptyDeploymentConfig := cloudflare.PagesProjectDeploymentConfigs{}
+	if !reflect.DeepEqual(project.DeploymentConfigs, emptyDeploymentConfig) {
+		deploymentConfigs := []map[string]interface{}{}
+		deploymentConfig := make(map[string]interface{})
+		emptyDeploymentEnviroment := cloudflare.PagesProjectDeploymentConfigEnvironment{}
+		if !reflect.DeepEqual(project.DeploymentConfigs.Preview, emptyDeploymentEnviroment) {
+			deploymentConfig["preview"] = parseDeployementConfig(project.DeploymentConfigs.Preview)
+		}
+
+		if !reflect.DeepEqual(project.DeploymentConfigs.Production, emptyDeploymentEnviroment) {
+			deploymentConfig["production"] = parseDeployementConfig(project.DeploymentConfigs.Production)
+		}
+		deploymentConfigs = append(deploymentConfigs, deploymentConfig)
+		d.Set("deployment_configs", deploymentConfigs)
+	}
 
 	return nil
 }
@@ -183,24 +286,24 @@ func resourceCloudflarePagesProjectCreate(ctx context.Context, d *schema.Resourc
 	accountID := d.Get("account_id").(string)
 	pageProject := buildPagesProject(d)
 
-	_, err := client.CreatePagesProject(ctx, accountID, pageProject)
+	project, err := client.CreatePagesProject(ctx, accountID, pageProject)
 	if err != nil {
 		return diag.FromErr(fmt.Errorf("error creating cloudflare pages project %q: %w", pageProject.Name, err))
 	}
 
+	d.SetId(project.Name)
 	return resourceCloudflarePagesProjectRead(ctx, d, meta)
 }
 
 func resourceCloudflarePagesProjectUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*cloudflare.API)
 	accountID := d.Get("account_id").(string)
-	projectName := d.Get("name").(string)
 
 	pageProject := buildPagesProject(d)
 
-	_, err := client.UpdatePagesProject(ctx, accountID, projectName, pageProject)
+	_, err := client.UpdatePagesProject(ctx, accountID, d.Id(), pageProject)
 	if err != nil {
-		return diag.FromErr(fmt.Errorf("error updating cloudflare pages project %q: %w", pageProject.Name, err))
+		return diag.FromErr(fmt.Errorf("error updating cloudflare pages project %q: %w", d.Id(), err))
 	}
 
 	return resourceCloudflarePagesProjectRead(ctx, d, meta)
@@ -209,11 +312,40 @@ func resourceCloudflarePagesProjectUpdate(ctx context.Context, d *schema.Resourc
 func resourceCloudflarePagesProjectDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	client := meta.(*cloudflare.API)
 	accountID := d.Get("account_id").(string)
-	projectName := d.Get("name").(string)
 
-	err := client.DeletePagesProject(ctx, accountID, projectName)
+	err := client.DeletePagesProject(ctx, accountID, d.Id())
 	if err != nil {
-		return diag.FromErr(fmt.Errorf("error deleting cloudflare pages project %q: %w", projectName, err))
+		return diag.FromErr(fmt.Errorf("error deleting cloudflare pages project %q: %w", d.Id(), err))
 	}
 	return nil
+}
+
+func resourceCloudflarePagesProjectImport(ctx context.Context, d *schema.ResourceData, meta interface{}) ([]*schema.ResourceData, error) {
+	client := meta.(*cloudflare.API)
+
+	// split the id so we can look up
+	idAttr := strings.SplitN(d.Id(), "/", 2)
+	var accountID string
+	var projectName string
+	if len(idAttr) == 2 {
+		accountID = idAttr[0]
+		projectName = idAttr[1]
+	} else {
+		return nil, fmt.Errorf("invalid id %q specified, should be in format \"accountID/project-name\" for import", d.Id())
+	}
+
+	project, err := client.PagesProject(ctx, accountID, projectName)
+	if err != nil {
+		return nil, fmt.Errorf("Unable to find record with ID %q: %w", d.Id(), err)
+	}
+
+	tflog.Info(ctx, fmt.Sprintf("Found project: %s", project.Name))
+
+	d.SetId(project.Name)
+	d.Set("name", project.Name)
+	d.Set("account_id", accountID)
+
+	resourceCloudflarePagesProjectRead(ctx, d, meta)
+
+	return []*schema.ResourceData{d}, nil
 }